+++ composeApp/src/commonMain/kotlin/io/github/sophon/fightingnerd/feat/home/usecase/EnsureMoveListIsCached.kt
@@ -1,38 +1,37 @@
 package io.github.sophon.fightingnerd.feat.home.usecase
 
+import io.github.aakira.napier.Napier
+import io.github.sophon.core.architecture.EmptyResult
 import io.github.sophon.core.architecture.Result
 import io.github.sophon.core.architecture.flatMap
-import io.github.sophon.core.architecture.map
-import io.github.sophon.core.featureConfig.model.Game
 import io.github.sophon.core.featureConfig.CoreFeatureRepo
-import io.github.sophon.core.wiki.model.WikiClient
+import io.github.sophon.core.featureConfig.model.Game
 import io.github.sophon.core.wiki.model.Character
-import io.github.sophon.core.wiki.model.Move
-import io.github.sophon.core.wiki.usecase.DownloadMoveListUseCase
+import io.github.sophon.core.wiki.model.WikiClient
 import io.github.sophon.fightingnerd.core.model.AppError
 import io.github.sophon.fightingnerd.core.util.mapWikiError
 
-internal class LoadMoveListUseCase(
+internal class EnsureMoveListIsCached(
     private val featureRepo: CoreFeatureRepo,
 ) {
     suspend fun invoke(
         game: Game,
-        characterQueryId: String,
-    ): Result<List<Move>, AppError> {
+        characterId: String,
+    ): EmptyResult<AppError> {
         val wikiClient = featureRepo.getWikiClientFor(game)
             ?: return Result.Error(AppError.WikiError(game.id))
-        val character = when (val result = wikiClient.fetchCharacter(characterQuery = characterQueryId)) {
+        val character = when (val result = wikiClient.fetchCharacter(characterQuery = characterId)) {
             is Result.Success -> result.data
             is Result.Error -> return Result.Error(AppError.WikiError(result.error.toString()))
         }
 
-        val result = wikiClient.fetchMoveList(characterQuery = characterQueryId)
+        val result = wikiClient.checkHasCachedMoves(characterId = character.id)
             .mapWikiError()
-            .flatMap { cachedMoveList ->
-                if (cachedMoveList.isEmpty()) {
-                    refreshMoveList(character, wikiClient)
+            .flatMap { hasCachedMoveList ->
+                if (hasCachedMoveList) {
+                    Result.Success(Unit)
                 } else {
-                    Result.Success(cachedMoveList)
+                    refreshMoveList(character, wikiClient)
                 }
             }
 
@@ -42,12 +41,11 @@ internal class LoadMoveListUseCase(
     private suspend fun refreshMoveList(
         character: Character,
         wikiClient: WikiClient
-    ): Result<List<Move>, AppError> {
+    ): EmptyResult<AppError> {
         val result = wikiClient.downloadMoveListFor(character)
             .mapWikiError()
             .flatMap { moveList ->
                 wikiClient.cacheMoveList(character, moveList)
-                    .map { moveList }
                     .mapWikiError()
             }
         return result