+++ feat/wikiDustLoop/src/commonMain/kotlin/io/github/sophon/wikidustloop/domain/DustLoopWikiClient.kt
@@ -14,6 +14,7 @@ import io.github.sophon.core.wiki.model.Character
 import io.github.sophon.core.wiki.model.Move
 import io.github.sophon.core.wiki.usecase.CacheCharacterListUseCase
 import io.github.sophon.core.wiki.usecase.CacheMoveListUseCase
+import io.github.sophon.core.wiki.usecase.CheckHasCachedMoveListUseCase
 import io.github.sophon.core.wiki.usecase.ClearCacheUseCase
 import io.github.sophon.core.wiki.usecase.DownloadCharacterListUseCase
 import io.github.sophon.core.wiki.usecase.DownloadMoveListUseCase
@@ -42,6 +43,7 @@ internal class DustLoopWikiClient(
     private val cacheMoveListUseCase: CacheMoveListUseCase,
     private val clearCacheUseCase: ClearCacheUseCase,
     private val fetchMoveListUseCase: FetchMoveListUseCase,
+    private val checkHasCachedMoveListUseCase: CheckHasCachedMoveListUseCase,
 
     private val getLastCacheInsertInstantUseCase: GetLastCacheInsertInstantUseCase,
     private val fetchMoveUseCase: FetchMoveUseCase,
@@ -92,6 +94,10 @@ internal class DustLoopWikiClient(
             .onError {
                 Napier.e(tag = TAG) { "downloadMoveList(${character.remoteQueryId}): $it" }
             }
+    }
+
+    override suspend fun checkHasCachedMoves(characterId: String): Result<Boolean, WikiError> {
+        return checkHasCachedMoveListUseCase.invoke(characterId)
     }
 
     override suspend fun cacheMoveList(