+++ feat/wikiSupercombo/src/commonMain/kotlin/io/github/sophon/wikiSuperCombo/integration/superComboModule.kt
@@ -11,6 +11,7 @@ import io.github.sophon.core.wiki.data.MoveListDB
 import io.github.sophon.core.wiki.model.WikiClient
 import io.github.sophon.core.wiki.usecase.CacheCharacterListUseCase
 import io.github.sophon.core.wiki.usecase.CacheMoveListUseCase
+import io.github.sophon.core.wiki.usecase.CheckHasCachedMoveListUseCase
 import io.github.sophon.core.wiki.usecase.ClearCacheUseCase
 import io.github.sophon.core.wiki.usecase.DownloadCharacterListUseCase
 import io.github.sophon.core.wiki.usecase.DownloadMoveListUseCase
@@ -58,15 +59,9 @@ fun superComboModule() = module {
                             .map { dto.toDomain(gameId, it) }
                     }
             },
-            cacheCharacterListUseCase = CacheCharacterListUseCase { characterList ->
-                characterListDB.insertCharacterList(characterList)
-            },
-            fetchCharacterUseCase = FetchCharacterUseCase { charName ->
-                characterListDB.fetchCharacterDataFor(charName)
-            },
-            fetchCharacterListUseCase = FetchCharacterListUseCase {
-                characterListDB.fetchCharacterList()
-            },
+            cacheCharacterListUseCase = CacheCharacterListUseCase(characterListDB::insertCharacterList),
+            fetchCharacterUseCase = FetchCharacterUseCase(characterListDB::fetchCharacterDataFor),
+            fetchCharacterListUseCase = FetchCharacterListUseCase(characterListDB::fetchCharacterList),
 
             downloadMoveListUseCase = DownloadMoveListUseCase { queryTable, characterData ->
                 source.downloadMoveList(queryTable.moves, characterData)
@@ -79,16 +74,10 @@ fun superComboModule() = module {
                 moveListDB.insertMoveList(game, character, moveList)
                     .asEmptyDataResult()
             },
-            fetchMoveUseCase = FetchMoveUseCase { charName, moveQuery ->
-                moveListDB.fetchMoveDataFor(charName, moveQuery)
-            },
-            fetchMoveListUseCase = FetchMoveListUseCase { charName ->
-                moveListDB.fetchMoveListFor(charName)
-            },
+            fetchMoveUseCase = FetchMoveUseCase(moveListDB::fetchMoveDataFor),
+            fetchMoveListUseCase = FetchMoveListUseCase(moveListDB::fetchMoveListFor),
 
-            getLastCacheInsertInstantUseCase = GetLastCacheInsertInstantUseCase {
-                moveListDB.getLastInsertTimeStamp()
-            },
+            getLastCacheInsertInstantUseCase = GetLastCacheInsertInstantUseCase(moveListDB::getLastInsertTimeStamp),
             clearCacheUseCase = ClearCacheUseCase {
                 val charResult = characterListDB.wipe()
                 val moveResult = moveListDB.wipe()
@@ -98,6 +87,7 @@ fun superComboModule() = module {
                     else -> Result.Success(Unit)
                 }
             },
+            checkHasCachedMoveListUseCase = CheckHasCachedMoveListUseCase(moveListDB::hasMovesCachedFor)
         )
     }
 }