+++ feat/xko/src/commonMain/kotlin/io/github/sophon/xko/domain/XkoWikiClient.kt
@@ -14,6 +14,7 @@ import io.github.sophon.core.wiki.model.Move
 import io.github.sophon.core.wiki.model.WikiClient
 import io.github.sophon.core.wiki.usecase.CacheCharacterListUseCase
 import io.github.sophon.core.wiki.usecase.CacheMoveListUseCase
+import io.github.sophon.core.wiki.usecase.CheckHasCachedMoveListUseCase
 import io.github.sophon.core.wiki.usecase.ClearCacheUseCase
 import io.github.sophon.core.wiki.usecase.DownloadOrFetchUseCase
 import io.github.sophon.core.wiki.usecase.FetchCharacterListUseCase
@@ -38,6 +39,7 @@ internal class XkoWikiClient(
     private val cacheMoveListUseCase: CacheMoveListUseCase,
     private val fetchMoveListUseCase: FetchMoveListUseCase,
     private val fetchMoveUseCase: FetchMoveUseCase,
+    private val checkHasCachedMoveListUseCase: CheckHasCachedMoveListUseCase,
 
     private val getLastCacheInsertInstantUseCase: GetLastCacheInsertInstantUseCase,
     private val clearCacheUseCase: ClearCacheUseCase,
@@ -88,6 +90,10 @@ internal class XkoWikiClient(
             .onError {
                 Napier.e(tag = TAG) { "downloadMoveList (${character.remoteQueryId}): $it" }
             }
+    }
+
+    override suspend fun checkHasCachedMoves(characterId: String): Result<Boolean, WikiError> {
+        return checkHasCachedMoveListUseCase.invoke(characterId)
     }
 
     override suspend fun cacheMoveList(